requirements.txt
pymupdf>=1.23.0
lxml>=4.9.0
pandas>=2.0.0
rapidfuzz>=3.0.0
streamlit>=1.33.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
openai>=1.42.0
